tests/wpt/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.rotate3d.z.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.transformation.rotate3d.z</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.rotate3d.z</h1>
<p class="desc">rotate3d() around the z axis results in the correct transformation matrix</p>


<script>
var t = async_test("rotate3d() around the z axis results in the correct transformation matrix");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

// angles are in radians, test something that is not a multiple of pi
const angle = 2;
const domMatrix = new DOMMatrix();
ctx.rotate3d(0, 0, angle);
domMatrix.rotateAxisAngleSelf(0, 0, 1, rad2deg(angle));
_assertMatricesApproxEqual(domMatrix, ctx.getTransform())
ctx.rotate3d(0, 0, angle);
domMatrix.rotateAxisAngleSelf(0, 0, 1, rad2deg(angle));
_assertMatricesApproxEqual(domMatrix, ctx.getTransform())
t.done();

});
</script>
